# infrastructure/terraform/modules/api-gateway/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Global Gaming Platform API
  description: |
    Unified API for the Global Gaming Platform - a tic-tac-toe gaming system with social media integration,
    real-time gameplay, and comprehensive leaderboard functionality.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT Bearer Token** (for user authentication)
       - Header: `Authorization: Bearer <token>`
       - Obtained through OAuth social login
    
    2. **API Key** (for service-to-service communication)
       - Header: `x-api-key: <api-key>`
       - Provided to authorized services
    
    ## Rate Limiting
    
    - **Authenticated users**: 1000 requests/minute
    - **Anonymous users**: 100 requests/minute
    - **Burst capacity**: 2000 requests
    
    ## Versioning
    
    API versioning is handled through the URL path:
    - Current version: v1
    - Format: `/v1/{resource}`
    
  version: 1.0.0
  contact:
    name: Gaming Platform API Support
    email: api-support@gaming-platform.com
  license:
    name: Proprietary
servers:
  - url: https://api.gaming-platform.com/v1
    description: Production server
  - url: https://api-staging.gaming-platform.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /leaderboard/global:
    get:
      tags:
        - Leaderboard
      summary: Get global leaderboard
      description: Returns the global leaderboard rankings
      operationId: getGlobalLeaderboard
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: Leaderboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /leaderboard/regional/{region}:
    get:
      tags:
        - Leaderboard
      summary: Get regional leaderboard
      description: Returns the leaderboard for a specific region
      operationId: getRegionalLeaderboard
      security: []
      parameters:
        - name: region
          in: path
          required: true
          schema:
            type: string
            enum: [europe, americas, asia, oceania, africa]
          description: Geographic region
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Regional leaderboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /leaderboard/user/{userId}:
    get:
      tags:
        - Leaderboard
      summary: Get user ranking
      description: Returns the ranking and statistics for a specific user
      operationId: getUserRanking
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User ranking retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRanking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ============================================================================
  # Support Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained through OAuth login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for service-to-service communication
...
  schemas:
...
    LeaderboardResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - rankings
            - total
          properties:
            rankings:
              type: array
              items:
                $ref: '#/components/schemas/LeaderboardEntry'
            total:
              type: integer
              description: Total number of ranked players
            limit:
              type: integer
            offset:
              type: integer
    LeaderboardEntry:
      type: object
      required:
        - rank
        - userId
        - username
        - gamesPlayed
        - gamesWon
        - winPercentage
      properties:
        rank:
          type: integer
          minimum: 1
        userId:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        gamesPlayed:
          type: integer
        gamesWon:
          type: integer
        gamesLost:
          type: integer
        gamesDrawn:
          type: integer
        winPercentage:
          type: number
          format: float
        currentStreak:
          type: integer
        bestStreak:
          type: integer
    UserRanking:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: '#/components/schemas/LeaderboardEntry'
            - type: object
              properties:
                globalRank:
                  type: integer
                regionalRank:
                  type: integer
                region:
                  type: string
    
    # Support Schemas
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: BAD_REQUEST
            message: Invalid request parameters
            timestamp: "2025-11-07T10:30:00Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: UNAUTHORIZED
            message: Authentication required
            timestamp: "2025-11-07T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: NOT_FOUND
            message: Resource not found
            timestamp: "2025-11-07T10:30:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: INTERNAL_ERROR
            message: An unexpected error occurred
            timestamp: "2025-11-07T10:30:00Z"